Prettier formatting for ./applications and ./interfaces

+++ packages/primitive-contracts/contracts/option/applications/factories/OptionFactory.sol
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: MIT
 
-
-
 pragma solidity ^0.6.2;
 
 /**
@@ -17,7 +15,7 @@ import { CloneLib } from "../../libraries/CloneLib.sol";
 import { IOptionFactory } from "../../interfaces/IOptionFactory.sol";
 
 contract OptionFactory is IOptionFactory, Ownable, NullCloneConstructor {
-    using SafeMath for uint256;
+    using SafeMath for uint;
 
     address public override optionTemplate;
 
@@ -32,40 +30,23 @@ contract OptionFactory is IOptionFactory, Ownable, NullCloneConstructor {
     function deploy(
         address underlyingToken,
         address strikeToken,
-        uint256 base,
-        uint256 quote,
-        uint256 expiry
+        uint base,
+        uint quote,
+        uint expiry
     ) external override onlyOwner returns (address option) {
         require(optionTemplate != address(0x0), "ERR_NO_DEPLOYED_TEMPLATE");
         bytes32 salt = keccak256(
-            abi.encodePacked(
-                OptionTemplateLib.OPTION_SALT(),
-                underlyingToken,
-                strikeToken,
-                base,
-                quote,
-                expiry
-            )
+            abi.encodePacked(OptionTemplateLib.OPTION_SALT(), underlyingToken, strikeToken, base, quote, expiry)
         );
-        option = CloneLib.create2Clone(optionTemplate, uint256(salt));
-        Option(option).initialize(
-            underlyingToken,
-            strikeToken,
-            base,
-            quote,
-            expiry
-        );
+        option = CloneLib.create2Clone(optionTemplate, uint(salt));
+        Option(option).initialize(underlyingToken, strikeToken, base, quote, expiry);
     }
 
     function kill(address option) external override onlyOwner {
         Option(option).kill();
     }
 
-    function initialize(address option, address redeem)
-        external
-        override
-        onlyOwner
-    {
+    function initialize(address option, address redeem) external override onlyOwner {
         Option(option).initRedeemToken(redeem);
     }
 }
