+++ packages/primitive-contracts/contracts/option/interfaces/ITrader.sol
@@ -1,15 +1,5 @@
 // SPDX-License-Identifier: MIT
 
-
-
-
-
-
-
-
-
-
-
 pragma solidity ^0.6.2;
 
 import { IOption } from "./IOption.sol";
@@ -17,43 +7,43 @@ import { IOption } from "./IOption.sol";
 interface ITrader {
     function safeMint(
         IOption optionToken,
-        uint256 mintQuantity,
+        uint mintQuantity,
         address receiver
-    ) external returns (uint256 inTokenU, uint256 outTokenR);
+    ) external returns (uint inTokenU, uint outTokenR);
 
     function safeExercise(
         IOption optionToken,
-        uint256 exerciseQuantity,
+        uint exerciseQuantity,
         address receiver
-    ) external returns (uint256 inTokenS, uint256 inOptions);
+    ) external returns (uint inTokenS, uint inOptions);
 
     function safeRedeem(
         IOption optionToken,
-        uint256 redeemQuantity,
+        uint redeemQuantity,
         address receiver
-    ) external returns (uint256 inRedeems);
+    ) external returns (uint inRedeems);
 
     function safeClose(
         IOption optionToken,
-        uint256 closeQuantity,
+        uint closeQuantity,
         address receiver
     )
         external
         returns (
-            uint256 inRedeems,
-            uint256 inOptions,
-            uint256 outUnderlyings
+            uint inRedeems,
+            uint inOptions,
+            uint outUnderlyings
         );
 
     function safeUnwind(
         IOption optionToken,
-        uint256 unwindQuantity,
+        uint unwindQuantity,
         address receiver
     )
         external
         returns (
-            uint256 inRedeems,
-            uint256 inOptions,
-            uint256 outUnderlyings
+            uint inRedeems,
+            uint inOptions,
+            uint outUnderlyings
         );
 }
